src/main/java/com/example/creationpattern/factorymethod/NotificationFactory.java
package com.example.creationpattern.factorymethod;

/**
 * @author Mahdi Sharifi
 * @since 9/22/22
 */
public class NotificationFactory {
    public Notification createNotification(String channel) {
        if (channel == null || channel.isEmpty())
            return null;
        return switch (channel) {
            case "SMS" -> new SMSNotification();
            case "EMAIL" -> new EmailNotification();
            case "PUSH" -> new PushNotification();
            default -> throw new IllegalArgumentException("Unknown channel " + channel);
        };
    }
}
